requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
markdown>=3.5.0
python-frontmatter>=1.0.0
